feat(lsp): TCP
