+++ taplo-lsp/Cargo.toml
@@ -42,7 +42,7 @@ tokio = { version = "0.3", features = [
     "io-util",
     "io-std",
     "sync",
-    "time"
+    "net"
 ] }
 tokio-compat-02 = { version = "0.1" }
 clap = "3.0.0-beta.2"
@@ -59,4 +59,4 @@ wasm-bindgen-test = "0.3"
 
 [[bin]]
 name = "taplo-lsp"
-path = "bin/lsp.rs"
+path = "bin/lsp/main.rs"
